+++ apps/api/src/course/course.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { CourseService } from './course.service';
 import { CourseController } from './course.controller';
-import { GlobalModule } from '../global/global.mudule';
+import { GlobalModule } from '../global/global.module';
 import { UserProgressService } from '../user-progress/user-progress.service';
 import { RankService } from '../rank/rank.service';
