+++ apps/api/src/game/game.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { GameService } from './game.service';
 import { GameController } from './game.controller';
-import { GlobalModule } from '../global/global.mudule';
+import { GlobalModule } from '../global/global.module';
 import { UserProgressModule } from '../user-progress/user-progress.module';
 import { CourseModule } from '../course/course.module';
